requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-socketio==5.10.0
websockets==11.0.3
python-dotenv==1.0.0
pydantic==2.4.2
loguru==0.7.2 